Correction - Models, Repositories, Routes, services -- Poids & Poids-Vignettes

+++ src/repositories/poids_repository.py
@@ -1,4 +1,4 @@
-from sqlmodel import Session
+from sqlalchemy.orm import Session
 from src.models.poids import Poids
 
 
@@ -11,7 +11,7 @@ class PoidsRepository:
         return list(db.query(Poids).all())
     
     def get_poids_by_id(self, db: Session, id: int):
-        return db.query(Poids).get(id)
+        return db.get(Poids,id)
     
     def create_poids(self, db: Session, donnees_poids: dict):
         poids = Poids(**donnees_poids)
@@ -21,7 +21,9 @@ class PoidsRepository:
         return poids
     
     def patch_poids(self, db: Session, id: int, donnees_poids: dict):
-        poids = db.query(Poids).get(id)
+        poids = db.get(Poids,id)
+        if not poids:
+            return None
         for key, value in donnees_poids.items():
             setattr(poids, key, value)
         db.commit()
@@ -29,7 +31,9 @@ class PoidsRepository:
         return poids
     
     def delete_poids(self, db: Session, id: int):
-        poids = db.query(Poids).get(id)
+        poids = db.get(Poids,id)
+        if not poids:
+            return None
         db.delete(poids)
         db.commit()
         return poids
