+++ src/repositories/poidsv_repository.py
@@ -3,15 +3,15 @@ from src.models.poidsv import Poidsv
 
 
 class PoidsvRepository:
-    """Repository class for managing poids-Vignettes operations in the database.
+    """Repository class for managing poids-vignettes operations in the database.
 
-    Receive a dictionary from service, save it to the database and return the poids-Vignettes model to service."""
+    Receive a dictionary from service, save it to the database and return the poids-vignettes model to service."""
     
     def get_all_poidsv(self, db: Session):
         return list(db.query(Poidsv).all())
     
     def get_poidsv_by_id(self, db: Session, id: int):
-        return db.query(Poidsv).get(id)
+        return db.get(Poidsv,id)
     
     def create_poidsv(self, db: Session, donnees_poidsv: dict):
         poidsv = Poidsv(**donnees_poidsv)
@@ -21,7 +21,9 @@ class PoidsvRepository:
         return poidsv
     
     def patch_poidsv(self, db: Session, id: int, donnees_poidsv: dict):
-        poidsv = db.query(Poidsv).get(id)
+        poidsv = db.get(Poidsv,id)
+        if not poidsv:
+            return None
         for key, value in donnees_poidsv.items():
             setattr(poidsv, key, value)
         db.commit()
@@ -29,7 +31,9 @@ class PoidsvRepository:
         return poidsv
     
     def delete_poidsv(self, db: Session, id: int):
-        poidsv = db.query(Poidsv).get(id)
+        poidsv = db.get(Poidsv,id)
+        if not poidsv:
+            return None
         db.delete(poidsv)
         db.commit()
         return poidsv
